requirements.txt
hydra-core
matplotlib
numpy
omegaconf
pandas
Pillow
pydicom
lightning
Requests
scikit-learn
scikit-image
torch
torchvision
tqdm
urllib3
wandb
transformers
timm
wilds
typing
open-clip-torch
